refactor: 공통 사용하는 클래스 common 패키지로 분리
- HttpStatus, HttpHeader 등 여러 곳에서 사용하는 클래스 common으로 분리
