+++ src/main/java/common/HttpHeader.java
@@ -1,4 +1,4 @@
-package webserver;
+package common;
 
 public enum HttpHeader {
     CONTENT_LENGTH("Content-Length"),
