+++ src/main/java/common/HttpResponse.java
@@ -1,4 +1,4 @@
-package webserver;
+package common;
 
 import java.util.HashMap;
 import java.util.Map;
