+++ src/main/java/common/HttpStatus.java
@@ -1,4 +1,4 @@
-package webserver;
+package common;
 
 public enum HttpStatus {
     OK(200, "OK"),
@@ -6,6 +6,7 @@ public enum HttpStatus {
     BAD_REQUEST(400, "Bad Request"),
     NOT_FOUND(404, "Not Found"),
     METHOD_NOT_ALLOWED(405, "Method Not Allowed"),
+    UNSUPPORTED_MEDIA_TYPE(415, "Unsupported Media Type"),
     INTERNAL_SERVER_ERROR(500, "Internal Server Error" );
 
     public final int code;
